- Used static classes for userController
- Codeclimate coverage reporting

- Removed Signin Controller

- Removed Signup Controller

[#168532824, #168532778]
