+++ .eslintrc
@@ -30,6 +30,7 @@
         ],
         "no-undef": 0,
         "import/prefer-default-export": 0,
+        "camelcase": 0
     }
  }
