+++ package.json
@@ -11,8 +11,7 @@
     "migrate": "NODE_ENV=test sequelize db:migrate",
     "rollback": "NODE_ENV=test sequelize db:migrate:undo:all",
     "coveralls-coverage": "coveralls < lcov.info",
-    "codeclimate-coverage": "codeclimate-test-reporter < lcov.info",
-    "coverage": "nyc npm test && npm run generate-lcov && npm run coveralls-coverage && npm run codeclimate-coverage"
+    "coverage": "nyc npm test && npm run generate-lcov && npm run coveralls-coverage"
   },
   "repository": {
     "type": "git",
