+++ routes/auth.js
@@ -1,12 +1,10 @@
 import { Router } from 'express';
 import { checkSignUp, verifyExistence, checkSignIn } from '../middlewares/authMiddleware';
-import { signupUser } from '../controllers/authentication/signupController';
-import { signinUser } from '../controllers/authentication/signinController';
+import userController from '../controllers/authentication/userController';
 
 const router = new Router();
 
-
-router.post('/signup', checkSignUp, verifyExistence, (req, res) => signupUser(req, res));
-router.post('/signin', checkSignIn, (req, res) => signinUser(req, res));
+router.post('/signup', checkSignUp, verifyExistence, (req, res) => userController.signupUser(req, res));
+router.post('/signin', checkSignIn, (req, res) => userController.signinUser(req, res));
 
 export default router;
